Fixed: SVY-3798 - Regression with dialogs from 6.1.2 SVY-3684 - Random "Error calling server" errors when closing dialogs in Firefox SVY-3804 - File upload callback method is not triggered in web client
Actual change:
- before, all div dialog actions would be scheduled to the parent (main)
frame, and executed on any request comming from either the main page of
any of it's child dialogs - directly from the response. There was an
exception, if the request was on a child and a show was supposed to
happen, then it didn't perform the actions right away (because show was
pretty restricted to running in main page). Also close browser side uses
setTimeout, breaking the order in which actions executed and leading to
these problems. Also any show that was called in JS after another dialog
close was called of any show called after another show would be delayed
to next main page request (but you could only delay the last show, the
rest being lost). This had 2 problems - you couldn't open multiple
dialogs from the same request and you couldn't be sure when the show was
called (breaking the browser side action execution order again).
- after this commit: all div dialog actions are scheduled to parent
(main frame), and executed on any request comming from either the main
page of any of it's child dialogs - but rather then doing it directly on
the response (that could close the same iframe => errors), these actions
are batched to be executed in sequence later on in the main page.
Because all actions are executed now in sequence but in parent/main page
(even show now works from children) close timeout was removed, being no
longer needed => the delayed dialog thing could be (and is) removed as
well.

The whole thing with executing all actions from children requests as
well started from an old case that only added 'close' to children as
well, breaking the order (some would execute from both children and
parents, some only from parent, some only from children), instead of
waiting for triggerAjaxUpdate... If this wasn't the case we could have
just removed the "delayed dialog" thing + scheduled all on main page
request and wait for that to get triggered...
Case: SVY-3798, SVY-3684, SVY-3804